Add files via upload
Since my personal library is divided between two countries, and books often transit from one side to the other, I decided it was time to organize things a bit.

The first step was to enter data in Excel. To minimize the time spend at this stage, after I decided on the desired information to be included (author, title, number of volumes, number of copies, language, whether it is read or not, the score that I would give to the books read (on a likert scale from 1 to 5, where 1 = "disappointing", and the location) I used the option "insert form" and I typed the information correctly, but chaotically.

When the above process was finished, I used the "Capitalize each word" function for the Author column and "Uppercase" for the Title column, both generated by Power Query. For the other columns, I used Find&Select Blanks because during data entry I did not waste time with repetitive information, and now the missing spaces had to be filled (for example, for the column containing the number of copies, I noted only the cases in which they were >1, the other spaces were filled automatically using the Find&Select function). 

In the end, I wanted to turn the score I gave to the books I read into stars, so I used the function =REPT()